Android420ToI420 - use ptrdiff_t for difference of u and v pointers
The difference was assigned to an int, causing a warning on Visual C.

BUG=662
TEST=tested with try bots.
R=devangelakos@google.com

Review-Url: https://codereview.chromium.org/2574373002 .

+++ source/convert.cc
@@ -1496,7 +1496,7 @@ int Android420ToI420(const uint8* src_y,
                      int width,
                      int height) {
   int y;
-  const int vu_off = src_v - src_u;
+  const ptrdiff_t vu_off = src_v - src_u;
   int halfwidth = (width + 1) >> 1;
   int halfheight = (height + 1) >> 1;
   if (!src_u || !src_v || !dst_u || !dst_v || width <= 0 || height == 0) {
